Add JSCS to improve coding standards
Someone asked me if the code they were writing was correct and standard enough.
As for the standard part I added a .jscsrc file that sets coding style to
Google's. This way if you install the linter-jscs package in Atom you will
immediately see what's not standard, and also fix it automatically (but I'm not
doing this on files I didn't write).

Needless to say, if we start using CoffeeScript we won't need a standards linter
anymore...