Turn double equals sign to triple equals
A best practice in JavaScript programming is to always use '===' instead of '=='.
This problem is not reported by JSCS but it usually is by JSHint. CoffeeScript
instead automatically turns double equals to triple equals when compiling to JS,
or simply allows you to use a 'is' keyword that compiles to '==='.