Rewrite completely validation
We dropped meteor-shower, we decided to stick with simple-form-data
instead. This lead us to use autoform, which does a great job at
creating inputs with almost no code, and some Bootstrap plugins such
as bootstrap-datepicker and bootstrap-switch.

NOTE: the Bootstrap-datepicker autoform field does not hold the date,
We still need to fix this!