Fix datepicker not holding value
I finally discovered what was happening: we were trying to save
dates as strings in the application! In order to make validation
work again I had to inhibit the 'change' event handler in the
apply.js file. We need to make an explicit conversion there or
in  the final 'submit' handler.