Fix recruiters visualization
I also created two recruiters so we can check if the recruiters filter
is working properly. It is not.

A query like 'phases.list.recruiter' will probably return the first recruiter
found in the phases list. What we want is the current one, that is the last one.
Maybe we could just save phases in descendent order, or fetch them in descendent
order?