Geolocation Autocomplete Feature
We are now looking for a better way to insert data in the form
and google maps API are an awesome start point.
Address field now works fine with GoogleMaps but the goal is to automatically
fill the entire fieldset and let the "aldeed"-"forwarder"-"i18n" & other
validation entities do their job with no troubles.