Create phases collection
There is no such thing as application or recruiting outcomes,
and also activity outcomes have nothing to do with applications.
We must deal with phases, which are much more simple than what we
came up with since now.

An application can be:
- new/unassigned
- in phase 1-3 (assigned)
- approved or rejected

This can be modeled through 6 phases. Maybe we could call them stages
or statuses, but in this case phases 1 through 3 actually keep their
ids, so they can be considered as the only phases the user knows about.